refactor: LV3 케이스 문 람다식 적용 객체 지향적 설계 시도 중
