+++ src/lv03/Calculator.java
@@ -74,48 +74,39 @@ public class Calculator {
 
     public void controlMainMenuByCommand(MainMenuCommandLine command) {
         switch (command) {
-            case CALCULATE:
-                // Arithmetic Calculator 계산 진행
+            case CALCULATE -> // Arithmetic Calculator 계산 진행
                 calculationMenu();
-                break;
-            case MEMORY:
-                // 메모리 메뉴로 이동
+
+            case MEMORY -> // 메모리 메뉴로 이동
                 showMemoryMenu();
-                break;
-            case EXIT:
-                // 시스템 종료
+
+            case EXIT -> // 시스템 종료
                 mainMenu.setSatae(false);
-                break;
         }
     }
 
     public void controlMemoryByCommand(MemoryMenuCommandLine command) {
         switch (command) {
-            case SHOW:
-                // 컬렉션 보여주기
+            case SHOW -> // 컬렉션 보여주기
                 memory.show();
-                break;
-            case DELETE_FIRST:
-                // 첫번째 삭제 후 컬렉션 보여주기
+
+            case DELETE_FIRST -> // 첫번째 삭제 후 컬렉션 보여주기
                 memory.deleteFirst();
-                break;
-            case FIND_BIGGER:
-                // 입력한 숫자보다 높은 값 컬렉션 보여주기
+
+            case FIND_BIGGER -> { // 입력한 숫자보다 높은 값 컬렉션 보여주기
                 try {
                     double target = Double.parseDouble(input.input());
                     memory.findBigger(target);
                 } catch (NumberFormatException e) {
                     e.getMessage();
                 }
-                break;
-            case CLEAR:
-                // 컬렉션 비우기
+            }
+
+            case CLEAR -> // 컬렉션 비우기
                 memory.clear();
-                break;
-            case BACK:
-                // 메인 메뉴로 돌아가기
+
+            case BACK -> // 메인 메뉴로 돌아가기
                 memoryMenu.setSatae(false);
-                break;
         }
     }
 }
