+++ src/lv03/calculation/AdditionCalculation.java
@@ -1,4 +1,4 @@
-package lv03;
+package lv03.calculation;
 
 public class AdditionCalculation implements Calculation {
     @Override
