+++ src/lv03/calculation/Calculation.java
@@ -1,4 +1,4 @@
-package lv03;
+package lv03.calculation;
 
 public interface Calculation {
     public double calculate(double num1, double num2);
