+++ src/lv03/calculation/DivideCalculation.java
@@ -1,4 +1,4 @@
-package lv03;
+package lv03.calculation;
 
 public class DivideCalculation implements Calculation {
     @Override
