+++ src/lv03/calculation/MinusCalculation.java
@@ -1,4 +1,4 @@
-package lv03;
+package lv03.calculation;
 
 public class MinusCalculation implements Calculation {
     @Override
