+++ src/lv03/calculation/ModuloCalculation.java
@@ -1,4 +1,4 @@
-package lv03;
+package lv03.calculation;
 
 public class ModuloCalculation implements Calculation {
     @Override
