+++ src/lv03/calculation/MultiplyCalculation.java
@@ -1,4 +1,4 @@
-package lv03;
+package lv03.calculation;
 
 public class MultiplyCalculation implements Calculation {
     @Override
